refactor: move slide view widget into page feature
